Update opus.c - 64b return type for opus_find_frame
Fixes support for files of > 2GB, prevents 32b rollover

+++ src/opus.c
@@ -349,7 +349,7 @@ out:
   return 0;
 }
 
-static int
+static off_t
 opus_find_frame(PerlIO *infile, char *file, int offset)
 {
   int frame_offset = -1;
